+++ flask_dance/consumer/storage/sqla.py
@@ -1,9 +1,8 @@
 from datetime import datetime
 
-from sqlalchemy import Column, Integer, String, DateTime
+from sqlalchemy import Column, Integer, String, DateTime, JSON
 from sqlalchemy.ext.mutable import MutableDict
 from sqlalchemy.ext.declarative import declared_attr
-from sqlalchemy_utils import JSONType
 from sqlalchemy.orm.exc import NoResultFound
 from flask_dance.utils import FakeCache, first
 from flask_dance.consumer.storage import BaseStorage
@@ -28,7 +27,7 @@ class OAuthConsumerMixin:
         an automatically generated datetime that indicates when
         the OAuth provider issued this token
     ``token``
-        a :class:`JSON <sqlalchemy_utils.types.json.JSONType>` field to store
+        a :class:`JSON <sqlalchemy.types.JSON>` field to store
         the actual token received from the OAuth provider
     """
 
@@ -39,7 +38,7 @@ class OAuthConsumerMixin:
     id = Column(Integer, primary_key=True)
     provider = Column(String(50), nullable=False)
     created_at = Column(DateTime, default=datetime.utcnow, nullable=False)
-    token = Column(MutableDict.as_mutable(JSONType), nullable=False)
+    token = Column(MutableDict.as_mutable(JSON), nullable=False)
 
     def __repr__(self):
         parts = []